requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
celery
redis
pytest-django
whitenoise
django-debug-toolbar
django-filter
PyJWT
dj-database-url
drf-nested-routers
drf-nested-routers
django-storages
boto3
drf-spectacular
django-celery-beat